Implement pre-event-handler hooks in the events system
These are somewhat general purpose, but was added for the
before REST handler hook to make sure that those event
handlers did not accidentally circumvent the intended
security mode of the route they were bound to. Now, any
event handler bound to a REST "before" event will be
required to specify an access level just like a normal
route handler, or else default to admin-only.

This commit also reinstates the small timeout for our
asyncore loop in the mock SMTP server.